Cat1_All_Endpoints/shodan_requirements.txt
requests==2.32.3
pymongo==4.14.0
dnspython==2.7.0
python-dotenv==1.0.0
